requirements.txt
# YouTube Data Fetching
google-api-python-client
youtube-transcript-api
isodate
python-dotenv

# ChromaDB and Vector Search
chromadb
tqdm

# Google Cloud (for embeddings)
google-cloud-aiplatform

# Optional: For GCP Vertex Search (if needed later)
google-cloud-storage
google-cloud-discoveryengine
